Other fixes:
- Renamed gpt3 file to ai.py and GPT3Client to OpenAiClient.
- Moved voice functions from youtube.py to voice.py.
